Added my own tut answers
