+++ Additional Tutorial 1/AddTut1.java
@@ -1,0 +1,6 @@
+public class AddTut1 {
+
+    public static void main(String[] args){
+        System.out.println("Hello\nAlexandra Abramov");
+    }
+}
