+++ Additional Tutorial 1/AddTut12.java
@@ -1,0 +1,10 @@
+public class AddTut12 {
+
+    public static void main (String[] args){
+        System.out.println(" +\"\"\"\"\"+ ");
+        System.out.println("[| o o |]");
+        System.out.println(" |  ^  |");
+        System.out.println(" | '_' |");
+        System.out.println(" +-----+");
+    }
+}
