+++ Additional Tutorial 1/AddTut2.java
@@ -1,0 +1,5 @@
+public class AddTut2 {
+    public static void main(String[] args){
+        System.out.println(74+36);
+    }
+}
