+++ Additional Tutorial 1/AddTut3.java
@@ -1,0 +1,6 @@
+public class AddTut3 {
+
+    public static void main (String [] args){
+        System.out.println(50/3);
+    }
+}
