+++ Additional Tutorial 1/AddTut4.java
@@ -1,0 +1,8 @@
+public class AddTut4 {
+    public static void main (String []args){
+        System.out.println(-5+8*6);
+        System.out.println((55+9)%9);
+        System.out.println(20+-3*5/8);
+        System.out.println(5+15/3*2-8%3);
+    }
+}
