+++ Additional Tutorial 1/AddTut6.java
@@ -1,0 +1,8 @@
+public class AddTut6 {
+    public static void main(String []args){
+        System.out.println("   J    a   v     v  a");
+        System.out.println("   J   a a   v   v  a a");
+        System.out.println("J  J  aaaaa   V V  aaaaa");
+        System.out.println(" JJ  a     a   V  a     a");
+    }
+}
