+++ Additional Tutorial 1/AddTut7.java
@@ -1,0 +1,5 @@
+public class AddTut7 {
+    public static void main(String[] args) {
+        System.out.println((25.5 * 3.5 - 3.5 * 3.5) / (40.5 - 4.5));
+    }
+}
