+++ Additional Tutorial 1/AddTut8.java
@@ -1,0 +1,6 @@
+public class AddTut8 {
+
+    public static void main(String []args){
+        System.out.println(4.0*(1-(1.0/3)+(1.0/5)-(1.0/7)+(1.0/9)-(1.0/11)));
+    }
+}
