+++ Additional Tutorial 1/AddTute10.java
@@ -1,0 +1,9 @@
+public class AddTute10 {
+
+    public static void main (String []args){
+        double width = 5.5;
+        double height = 8.5;
+        System.out.println("Area is "+width*height);
+        System.out.println("Perimeter is "+ 2*(width+height) );
+    }
+}
