+++ Additional Tutorial 1/AddTute11.java
@@ -1,0 +1,6 @@
+public class AddTute11 {
+
+    public static void main (String []args){
+        System.out.println("* * * * * * ==================================\n* * * * * * ==================================\n* * * * * * ==================================\n* * * * * * ==================================\n* * * * * * ==================================\n* * * * *  ==================================\n* * * * * * ==================================\n* * * * * * ==================================\n* * * * * * ==================================\n==============================================\n==============================================\n==============================================\n==============================================\n==============================================\n==============================================");
+    }
+}
