--- Additional Tutorial 1/Question1.java
@@ -1,9 +1,0 @@
-//Additional Tutorial 1 - Q1
-
-public class Question1{
-	
-	public static void main(String[] args){
-		System.out.println("Hello\nAlexandra Abramov");
-	}
-	
-}
