--- Additional Tutorial 1/Question10.java
@@ -1,10 +1,0 @@
-//Additional Tutorial - Q10
-
-public class Question10{
-	public static void main (String []args){
-		double width = 5.5;
-		double height = 8.5;
-		System.out.println("Area is "+Double.toString(width)+"*"+Double.toString(height)+"="+(width*height));
-		System.out.println("Perimeter is 2*("+Double.toString(width)+"+"+Double.toString(height)+")="+(2*(width+height)));
-	}
-}
