--- Additional Tutorial 1/Question11.java
@@ -1,7 +1,0 @@
-//Additional Tutorial - Q11
-
-public class Question11{
-	public static void main (String []args){
-		System.out.println("* * * * * * ==================================\n* * * * *  ==================================\n* * * * * * ==================================\n* * * * *  ==================================\n* * * * * * ==================================\n* * * * *  ==================================\n* * * * * * ==================================\n* * * * *  ==================================\n* * * * * * ==================================\n==============================================\n==============================================\n==============================================\n==============================================\n==============================================\n==============================================");
-	}
-}
