--- Additional Tutorial 1/Question12.java
@@ -1,11 +1,0 @@
-//Additional Tutorial 1 - Q12
-
-public class Question12{
-	public static void main (String[] args){
-		System.out.println(" +\"\"\"\"\"+ ");
-		System.out.println("[| o o |]");
-		System.out.println(" |  ^  |");
-		System.out.println(" | '_' |");
-		System.out.println(" +-----+");
-	}
-}
