--- Additional Tutorial 1/Question2.java
@@ -1,7 +1,0 @@
-//Additional Tutorial 1 - Q2
-
-public class Question2{
-	public static void main(String[] args){
-		System.out.println(74+36);
-	}
-}
