--- Additional Tutorial 1/Question3.java
@@ -1,7 +1,0 @@
-//Additional Tutorial - Q3
-
-public class Question3{
-	public static void main (String [] args){
-		System.out.println(Math.round(50/3));
-	}
-}
