--- Additional Tutorial 1/Question4.java
@@ -1,10 +1,0 @@
-//Additional Tutorial - Q4
-
-public class Question4{
-	public static void main (String []args){
-		System.out.println(-5+8*6);
-		System.out.println((55+9)%9);
-		System.out.println(20+-3*5/8);
-		System.out.println(5+15/3*2-8%3);
-	}
-}
