--- Additional Tutorial 1/Question6.java
@@ -1,10 +1,0 @@
-//Additional Tutorial - Q6
-
-public class Question6{
-	public static void main(String []args){
-		System.out.println("   J    a   v     v  a");
-		System.out.println("   J   a a   v   v  a a");
-		System.out.println("J  J  aaaaa   V V  aaaaa");
-		System.out.println(" JJ  a     a   V  a     a");
-	}
-}
