--- Additional Tutorial 1/Question7.java
@@ -1,6 +1,0 @@
-//Additional Tutorial - Q7
-
-public class Question7{
-	public static void main(String []args){
-		System.out.println((25.5*3.5-3.5*3.5)/(40.5-4.5));
-	}
