--- Additional Tutorial 1/Question9.java
@@ -1,9 +1,0 @@
-//Additional Tutorial - Q9
-
-public class Question9{
-	public static void main(String []args){
-		double r = 7.5;
-		System.out.println("Perimeter is = "+(2*3.14*r));
-		System.out.println("Area is = "+(3.14*(Math.pow(r,2))));
-	}
-}
